6 - Sistemas Distribuidos/Atividades Avaliativas/05 - RPC/client/requirements.txt
grpcio==1.54.0
grpcio-tools==1.54.0
protobuf==3.20.3